Moved canvas-related buttons into image utilities component
All the functionality still works the same since the canvas ref is passed through.
I'm considering refactoring again such that the CanvasDraw is it's own component again, however that may cause complications with the different useState hooks. It shouldn't, but that refactor isn't the main focus at the moment.
